Encapsulating ConfettiType
